moved feature handling to parent class
