+++ src/compas_timber/elements/timber.py
@@ -1,4 +1,5 @@
 from compas_model.elements import Element
+from compas_model.elements import reset_computed
 
 
 class TimberElement(Element):
@@ -19,8 +20,15 @@ class TimberElement(Element):
 
     """
 
-    def __init__(self, *args, **kwargs):
-        super(TimberElement, self).__init__(*args, **kwargs)
+    @property
+    def __data__(self):
+        data = super(TimberElement, self).__data__
+        data["features"] = [f for f in self.features if not f.is_joinery]  # type: ignore
+        return data
+
+    def __init__(self, features=None, **kwargs):
+        super(TimberElement, self).__init__(**kwargs)
+        self._features = features or []
         self.debug_info = []
 
     @property
@@ -43,8 +51,14 @@ class TimberElement(Element):
     def is_fastener(self):
         return False
 
-    def remove_features(self):
-        pass
+    @property
+    def features(self):
+        # type: () -> list[Feature]
+        return self._features
+
+    @features.setter
+    def features(self, features):
+        self._features = features
 
     def remove_blank_extension(self):
         pass
@@ -54,3 +68,36 @@ class TimberElement(Element):
         self.remove_features()
         self.remove_blank_extension()
         self.debug_info = []
+
+    @reset_computed
+    def add_features(self, features):
+        # type: (Feature | list[Feature]) -> None
+        """Adds one or more features to the beam.
+
+        Parameters
+        ----------
+        features : :class:`~compas_timber.parts.Feature` | list(:class:`~compas_timber.parts.Feature`)
+            The feature to be added.
+
+        """
+        if not isinstance(features, list):
+            features = [features]
+        self._features.extend(features)  # type: ignore
+
+    @reset_computed
+    def remove_features(self, features=None):
+        # type: (None | Feature | list[Feature]) -> None
+        """Removes a feature from the beam.
+
+        Parameters
+        ----------
+        feature : :class:`~compas_timber.parts.Feature` | list(:class:`~compas_timber.parts.Feature`)
+            The feature to be removed. If None, all features will be removed.
+
+        """
+        if features is None:
+            self._features = []
+        else:
+            if not isinstance(features, list):
+                features = [features]
+            self._features = [f for f in self._features if f not in features]
